trying it as a composite
